requirements.txt
asgiref==3.8.1
Django==4.2.25
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg2-binary>=2.9.10
python-dotenv==0.21.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.7.0
drf-yasg==1.21.8
